+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [0.0.7] - 2024-05-30
+### Improved
+- table fetch method
+- table reload method
+
 ## [0.0.6] - 2024-05-28
 ### Added
 - add animations in RouterView